Get locale information from the server rather than loading it from the theme
This fixes the bug where clicking on a locale link in the locale
selector will only select the new locale sometimes. This happened
because the client would sometimes get desynced from the server and have
the wrong session link. This new system will only use the session link
from the server.